requirements.txt
pandas==1.4.3
requests
flask
flask-cors
camelot-py==0.11.0
ghostscript==0.7
